マイグレーションファイルのに条件文追加

+++ db/migrate/20250318034236_remove_order_id_id_from_delivery_addresses.rb
@@ -1,5 +1,7 @@
 class RemoveOrderIdIdFromDeliveryAddresses < ActiveRecord::Migration[7.1]
   def change
-    remove_column :delivery_addresses, :order_id_id
+    if column_exists?(:delivery_addresses, :order_id_id)
+      remove_column :delivery_addresses, :order_id_id
+    end
   end
 end
